Java Doc comments and pre conditions
Adding pre conditions (double check them and test)

+++ CT5005_booking/src/accounts/AttendeeManager.java
@@ -1,3 +1,9 @@
+/**
+ * @author Joshua Preece
+ * @description Class handles all database communication for the attendees table and functions 
+ * @version 0.6
+ */
+
 package accounts;
 
 import java.sql.ResultSet;
@@ -23,14 +29,19 @@ public class AttendeeManager implements IDatabaseFunctions {
 		
 		try {
 
-			if (att.getAge() <= 12) {
+			// Check if the database does not have the maximum number of attendees (pre-condition)
+			if (DatabaseManager.count_items("attendees") <= Festival.MAX_ATTENDEES) {
 				
-				// do something if attendee is under 12
+				if (att.getAge() <= 12) {
+					
+					// do something if attendee is under 12
+					
+				} else {
 				
-			} else {
-			
-				add_entry(att);
-			
+					add_entry(att);
+				
+				}
+				
 			}
 			
 			System.out.println("Create attendee successful...");
@@ -45,7 +56,16 @@ public class AttendeeManager implements IDatabaseFunctions {
 		
 		try {
 			
-			DatabaseManager.print_results("Attendee Search Result", DatabaseManager.search_database("attendees", column, data));
+			// Check to see if the database is not empty (pre-condition)
+			if (DatabaseManager.count_items("attendees") < 0) {
+				
+				DatabaseManager.print_results("Attendee Search Result", DatabaseManager.search_database("attendees", column, data));
+			
+			} else {
+				
+				ErrorLog.printInfo("No items found in the database, because the database is empty!");
+				
+			}
 			
 		} catch (SQLException e) {
 			ErrorLog.printError("Search for attendee failed!\n" + e.getMessage(), ErrorLog.SEVERITY_MEDIUM);
@@ -58,9 +78,17 @@ public class AttendeeManager implements IDatabaseFunctions {
 		
 		try {
 			
-			remove_entry(att.getRef());
+			// Check if the attendee exists in the database before removing (pre-condition)
+			if (DatabaseManager.does_entry_exist("attendess", "ref", att.getRef())) {
+				
+				remove_entry(att.getRef());
+				System.out.println("Remove attemdee successful...");
 			
-			System.out.println("Remove attemdee successful...");
+			} else {
+				
+				System.out.println("Attendee does not exist. Please check the Ref!");
+				
+			}
 			
 		} catch (SQLException e) {
 			ErrorLog.printError("Remove attendee failed!\n" + e.getMessage(), ErrorLog.SEVERITY_MEDIUM);
@@ -68,6 +96,9 @@ public class AttendeeManager implements IDatabaseFunctions {
 		
 	}
 
+	/**
+	 * @Pre-Conditions The attendees must not have the maximum number of attendees in this case 4000.
+	 */
 	@Override
 	public boolean add_entry(Object data) throws SQLException {
 		
@@ -94,28 +125,44 @@ public class AttendeeManager implements IDatabaseFunctions {
 	public void remove_entry(String ref) throws SQLException {
 
 		Statement stat = DatabaseManager.getConnection().createStatement();
+		
+		// delete entry only if it exists
+		if (DatabaseManager.does_entry_exist("attendees", "ref", ref)) {
 			
 		stat.execute("DELETE FROM attendees WHERE ref=" + ref);
 
+		}
+		
 		stat.close();
 		
 	}
 	
+	/**
+	 * @Pre-Conditions Check if the booking assigned to the attendee, 
+	 * ensure it does not already have the maximum number of attendees assigned.
+	 */
 	@Override
 	public void update_entry(Object data) throws SQLException {
 		
 		Attendee att = (Attendee)data;
-		
 		Statement stat = DatabaseManager.getConnection().createStatement();
 		
-		Booking b = new Booking();
-		att.setBooking(b.getRef());
-		att.setBooking("1");
-		att.toString();
+		// validate that a booking does not already have the max number of attendees
+		if (DatabaseManager.count_specific_items("attendees", "booking", att.getBooking()) <= 4) {
 		
-		stat.executeUpdate("UPDATE attendees SET name='" + att.getName() + "', age='"
-				+ Integer.toString(att.getAge()) + "', email_address='" + att.getEmailAddress() 
-				+ "', booking='" + att.getBooking() + "' WHERE ref=" + att.getRef());
+			Booking b = new Booking();
+			att.setBooking(b.getRef());
+			att.setBooking("1");
+			att.toString();
+			
+			stat.executeUpdate("UPDATE attendees SET name='" + att.getName() + "', age='"
+					+ Integer.toString(att.getAge()) + "', email_address='" + att.getEmailAddress() 
+					+ "', booking='" + att.getBooking() + "' WHERE ref=" + att.getRef());
+		
+		} else {
+			
+			
+		}
 		
 		stat.close();
 		
@@ -124,6 +171,7 @@ public class AttendeeManager implements IDatabaseFunctions {
 	@Override
 	public void create_table() throws SQLException {
 		
+		// Ensure table does not exist before creating it
 		if (DatabaseManager.does_table_exist("attendees")) {
 			
 			ErrorLog.printInfo("Table 'attendees' already exists");
@@ -149,6 +197,7 @@ public class AttendeeManager implements IDatabaseFunctions {
 	@Override
 	public void drop_table() throws SQLException {
 		
+		// Ensure that the database exists before
 		if (DatabaseManager.does_table_exist("attendees")) {
 			
 			Statement stat = DatabaseManager.getConnection().createStatement();
